Move Element content related methods into modules.
To clean up the code two new modules now hold methods concerning contents and essences for the element model.
